feat(common): move `get_registered_pool_user` utility
- Introduced `get_registered_pool_user` in `common.py` to create and
  register pool users.
- Removed duplicate implementation of `get_registered_pool_user` from
  `conway_common.py`.
- Updated references in tests to use the new utility in `common.py`.

This change improves code reuse and simplifies pool user creation.
