+++ cardano_node_tests/tests/tests_conway/test_committee.py
@@ -62,9 +62,9 @@ def pool_user(
     """Create a pool user."""
     key = helpers.get_current_line_str()
     name_template = common.get_test_id(cluster)
-    return conway_common.get_registered_pool_user(
-        cluster_manager=cluster_manager,
+    return common.get_registered_pool_user(
         name_template=name_template,
+        cluster_manager=cluster_manager,
         cluster_obj=cluster,
         caching_key=key,
     )
@@ -79,9 +79,9 @@ def pool_user_lg(
     cluster, __ = cluster_lock_governance
     key = helpers.get_current_line_str()
     name_template = common.get_test_id(cluster)
-    return conway_common.get_registered_pool_user(
-        cluster_manager=cluster_manager,
+    return common.get_registered_pool_user(
         name_template=name_template,
+        cluster_manager=cluster_manager,
         cluster_obj=cluster,
         caching_key=key,
     )
@@ -96,9 +96,9 @@ def pool_user_ug(
     cluster, __ = cluster_use_governance
     key = helpers.get_current_line_str()
     name_template = common.get_test_id(cluster)
-    return conway_common.get_registered_pool_user(
-        cluster_manager=cluster_manager,
+    return common.get_registered_pool_user(
         name_template=name_template,
+        cluster_manager=cluster_manager,
         cluster_obj=cluster,
         caching_key=key,
     )
