+++ cardano_node_tests/tests/tests_conway/test_guardrails.py
@@ -70,9 +70,9 @@ def pool_user(
     """Create a pool user for "lock governance"."""
     cluster, __ = cluster_guardrails
     name_template = common.get_test_id(cluster)
-    return conway_common.get_registered_pool_user(
-        cluster_manager=cluster_manager,
+    return common.get_registered_pool_user(
         name_template=name_template,
+        cluster_manager=cluster_manager,
         cluster_obj=cluster,
     )
