+++ cardano_node_tests/tests/tests_conway/test_info.py
@@ -37,9 +37,9 @@ def pool_user_ug(
     cluster, __ = cluster_use_governance
     key = helpers.get_current_line_str()
     name_template = common.get_test_id(cluster)
-    return conway_common.get_registered_pool_user(
-        cluster_manager=cluster_manager,
+    return common.get_registered_pool_user(
         name_template=name_template,
+        cluster_manager=cluster_manager,
         cluster_obj=cluster,
         caching_key=key,
     )
